+++ app/models/tools/home.rb
@@ -1,0 +1,5 @@
+module Tools
+  # The homepage of the application.
+  class Home < BitCore::Tool
+  end
+end