+++ db/migrate/20150130160818_add_type_to_tools.bit_core.rb
@@ -1,0 +1,6 @@
+# This migration comes from bit_core (originally 20150130155627)
+class AddTypeToTools < ActiveRecord::Migration
+  def change
+    add_column :bit_core_tools, :type, :string
+  end
+end