+++ db/migrate/20150130173324_convert_home_tool_type.rb
@@ -1,0 +1,7 @@
+class ConvertHomeToolType < ActiveRecord::Migration
+  def change
+    BitCore::Tool.where(title: "home").each do |tool|
+      tool.update type: "Tools::Home"
+    end
+  end
+end