+++ db/migrate/20150130185412_convert_learn_tool_type.rb
@@ -1,0 +1,7 @@
+class ConvertLearnToolType < ActiveRecord::Migration
+  def change
+    BitCore::Tool.where(title: "LEARN").each do |tool|
+      tool.update type: "Tools::Learn"
+    end
+  end
+end