+++ spec/features/participant/activity_tracker_spec.rb
@@ -48,10 +48,15 @@ feature "activity tracker", type: :feature do
 
       find("#copy_2").trigger("click")
 
+      find('button#submit_activities[type="submit"]').click
+
+      expect(page).to have_css("div.has-error")
+
       find("#activity_type_3").trigger("click")
       fill_in("activity_type_3", with: "ate bad cheeseburgers")
       choose_rating("pleasure_3", 0)
       choose_rating("accomplishment_3", 1)
+
       find('button#submit_activities[type="submit"]').click
 
       expect(page).to have_text("ate cheeseburgers")
@@ -91,9 +96,6 @@ feature "activity tracker", type: :feature do
       awake_period = participant1.awake_periods.build(start_time: seven_am, end_time: eight_pm)
       awake_period.save!
 
-      within "#sc-hamburger-menu" do
-        click_on "Home"
-      end
       visit "/navigator/contexts/DO"
       within ".container .left.list-group" do
         click_on "#1 Awareness"